% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{write_cnm}
\alias{write_cnm}
\title{Serialise a .cnm object to a file}
\usage{
write_cnm(cnm, file)
}
\arguments{
\item{cnm}{A list from [build_cnm()] / [cinema_cnm()].}

\item{file}{Output path.}
}
\value{
The output path, invisibly.
}
\description{
Serialise a .cnm object to a file
}
